chore(gpui): tighten comments + split diff_view/mod.rs
Two threads bundled (process slip — should have \`jj new\`-ed
between the comment cleanup and the mod.rs split):

1. Comments collapsed to one-liners per AGENTS.md principle 6:
   selection_overlay, DiffSelection, col_range_for, PanelBoundsSlot,
   bounds_capture, parse_command, mono_advance, sbs_body inline,
   SwiftUI rectArray override.

2. Split diff/diff_view/mod.rs (was 155 lines) into:
   - state.rs — DiffViewMode, DetailMode, DiffViewState, FindState
   - render.rs — diff_view() dispatch fn
   - mod.rs (11 lines) — pub mod + pub use re-exports
   Mirrors the log/ layout. Callers like log/detail/mod.rs keep
   working unchanged via the re-export.

Also pulled \`Pixels\` and \`UniformList\` into sbs_body.rs's
gpui::{} use list instead of qualifying inline.

Build + clippy + tests clean. Same content as the just-pushed
d244563c (which got rewritten into this commit during the split
work) — net effect on the branch is identical.
