+++ shell/gpui/src/diff/diff_view/state.rs
@@ -1,0 +1,35 @@
+use jayjay_core::DiffHunk;
+use jayjay_core::diff::FileDiff;
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum DiffViewMode {
+    Unified,
+    SideBySide,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum DetailMode {
+    Diff,
+    Annotate,
+}
+
+/// Pure-data inputs for the diff/annotate body.
+pub struct DiffViewState<'a> {
+    pub hunk: Option<&'a DiffHunk>,
+    pub file_diff: Option<&'a FileDiff>,
+    pub view_mode: DiffViewMode,
+    pub detail_mode: DetailMode,
+    pub annotate_lines: Option<std::sync::Arc<Vec<jayjay_core::AnnotationLine>>>,
+    pub loading_annotate: bool,
+    pub path_just_copied: bool,
+    pub unified_bounds: crate::log::PanelBoundsSlot,
+    pub sbs_old_bounds: crate::log::PanelBoundsSlot,
+    pub sbs_new_bounds: crate::log::PanelBoundsSlot,
+}
+
+/// Find-in-diff state.
+pub struct FindState<'a> {
+    pub query: Option<&'a str>,
+    pub match_count: usize,
+    pub match_current: usize,
+}
